Simpify media naming,
Every item will get the unique name, meaning it will never be overwritten.
Items will remain in the cloud so the versioning of the objects work.
This might not be useful when using provider for something else then eZ
content.
Todo: create way to purge items that are no more connected to any eZ
objects from Cloudinary.